veil updates
